[Rename] `registration.ts` to `registrant.ts`
